refactor : 모델값들 변경에 의한 BaseballGame 클래스 리팩토링 (그리고 controller로 이동)
